+++ hyperdbg/hprdbgctrl/code/debugger/core/debugger.cpp
@@ -381,6 +381,11 @@ ShowErrorMessage(UINT32 Error)
                      Error);
         break;
 
+    case DEBUGGER_ERROR_UNABLE_TO_KILL_THE_PROCESS_DOES_NOT_EXISTS:
+        ShowMessages("err, process does not exists (already terminated?) (%x)\n",
+                     Error);
+        break;
+
     default:
         ShowMessages("err, error not found (%x)\n",
                      Error);